Fixed the Merge

+++ backend/experiments.py
@@ -40,11 +40,17 @@ from backend.db import DBManager
 # sys.path.append("Classes")
 # sys.path.insert(1, '/Classes/')
 from backend import MLA
-from backend import Validation
+from backend import MLA_Validation
+from backend import DLANN_Validation
 from backend.Classes.PreoptimizationFiles import Preoptimization
 from backend.Classes.PreoptimizationFiles import Standardization
 from backend.Classes.NeuralNetworkFiles import NeuralNetwork
 from backend.Classes.NeuralNetworkFiles import Core
+from backend.Classes.NeuralNetworkFiles import Compiler
+from backend.Classes.NeuralNetworkFiles import Validation as NN_Validation
+from backend.Classes.NeuralNetworkFiles import Callbacks
+
+from keras.utils import to_categorical 
 
 from sklearn.model_selection import KFold
 
@@ -55,8 +61,6 @@ logger = logging.getLogger()
 
 # should be modified to work with a db or class if necessary.
 # Otherwise, the choices being pre-available on page load will slightly speed the process.
-
-
 @bp.before_app_request
 def load_possible_experiments():
     list_Algorithms = MLA.getMLAs()
@@ -66,25 +70,25 @@ def load_possible_experiments():
     # list for Neural Network options
     list_Layer_Categories = NeuralNetwork.getLayer_Categories()
     list_Core = Core.getCore()
+    # list for Callback options
+    list_Callbacks = Callbacks.getCallbacks()
 
-    # ----------------Data Types-------------------------->
+    #----------------Data Types-------------------------->
     g.data_types = ["Tabular", "Image", "Text", "Signal"]
     g.data_types_Name = ["Tabular", "Image", "Text", "Signal"]
     g.data_types_Info = [["Tabular is data that is displayed in columns or tables and does not pertain to the other data types."],
-                         ["Image is data that is usually 3-dimentional corresponding to the three color spectrom, RGB"],
-                         ["Text is data containing textual words or sentences."],
-                         ["Signal is data containing time-series data corresponding to a signal format."]]
+                      ["Image is data that is usually 3-dimentional corresponding to the three color spectrom, RGB"],
+                      ["Text is data containing textual words or sentences."],
+                      ["Signal is data containing time-series data corresponding to a signal format."]]
     g.Data_Types = zip(g.data_types, g.data_types_Name, g.data_types_Info)
 
-    # ----------------new Cycon--------------------------->
-    # First choice is between various categories of ML
+    #----------------new Cycon--------------------------->
+    # First choice is between various categories of ML 
     g.section_Method = ["MLA", "DLANN"]
-    g.section_Method_Display_Name = [
-        "Machine Learning Algorithm (MLA)", "Deep Learning Artifical Neural Networks (DLANN)"]
+    g.section_Method_Display_Name = ["Machine Learning Algorithm (MLA)", "Deep Learning Artifical Neural Networks (DLANN)"]
     g.section_Info = [["Machine learning algorithms are mathematical model mapping methods. They are used to learn patterns embedded in the existing training dataset in order to perform pattern recognition, classification, and prediction.\n\nCurrently, the only algorithms available on cycon is under the classification objective. Other objectives that will be added later include clustering and regression."],
                       ["DLANN info"]]
-    g.Methodologies = zip(
-        g.section_Method, g.section_Method_Display_Name, g.section_Info)
+    g.Methodologies = zip(g.section_Method, g.section_Method_Display_Name, g.section_Info)
 
     # Obtain the selection of Algorithm Names and Definition.
     algorithm_Names = []
@@ -99,7 +103,7 @@ def load_possible_experiments():
     g.section_Info = algorithm_Definition
     g.Algorithms = zip(g.section_algorithm, g.section_Info)
 
-    # ----------------Preoptimiztion-------------------------->
+    #----------------Preoptimiztion-------------------------->
     # create list of options to choose a category of preoptimization options.
     preopt_Category_Names = []
     preopt_Category_Display_Names = []
@@ -111,9 +115,8 @@ def load_possible_experiments():
         preopt_Category_Definition.append(preopt_Category.getDefinition())
     g.section_preopt_Category_Names = preopt_Category_Names
     g.section_preopt_Category_Display_Names = preopt_Category_Display_Names
-    g.section_preopt_Category_Info = preopt_Category_Definition
-    g.preopt_Categories = zip(g.section_preopt_Category_Names,
-                              g.section_preopt_Category_Display_Names, g.section_preopt_Category_Info)
+    g.section_preopt_Category_Info =  preopt_Category_Definition
+    g.preopt_Categories = zip(g.section_preopt_Category_Names, g.section_preopt_Category_Display_Names, g.section_preopt_Category_Info)
 
     # create list of options for the first category of preoprtimization (I.E. Standardization)
     preopt_Names = []
@@ -126,11 +129,10 @@ def load_possible_experiments():
         preopt_Definition.append(preopt.getDefinition())
     g.section_preopt_Names = preopt_Names
     g.section_preopt_Display_Names = preopt_Display_Names
-    g.section_preopt_Info = preopt_Definition
-    g.preopts = zip(g.section_preopt_Names,
-                    g.section_preopt_Display_Names, g.section_preopt_Info)
+    g.section_preopt_Info =  preopt_Definition
+    g.preopts = zip(g.section_preopt_Names, g.section_preopt_Display_Names, g.section_preopt_Info)
 
-    # -------------------Neural Network Layers----------------------->
+    #-------------------Neural Network Layers----------------------->
     # create list of options to choose a category of layer options.
     layer_Category_Names = []
     layer_Category_Display_Names = []
@@ -142,9 +144,8 @@ def load_possible_experiments():
         layer_Category_Definition.append(layer_Category.getDefinition())
     g.layer_Category_Names = layer_Category_Names
     g.layer_Category_Display_Names = layer_Category_Display_Names
-    g.layer_Category_Info = layer_Category_Definition
-    g.NN_layer_Categories = zip(
-        g.layer_Category_Names, g.layer_Category_Display_Names, g.layer_Category_Info)
+    g.layer_Category_Info =  layer_Category_Definition
+    g.NN_layer_Categories = zip(g.layer_Category_Names, g.layer_Category_Display_Names, g.layer_Category_Info)
 
     # create list of options for the first category of preoprtimization (I.E. Core)
     layer_Names = []
@@ -157,9 +158,23 @@ def load_possible_experiments():
         layer_Definition.append(layer.getDefinition())
     g.layer_Names = layer_Names
     g.layer_Display_Names = layer_Display_Names
-    g.layer_Info = layer_Definition
+    g.layer_Info =  layer_Definition
     g.layers = zip(g.layer_Names, g.layer_Display_Names, g.layer_Info)
 
+    #-------------------Callbacks----------------------->
+    # create list of options to choose a callback options.
+    callback_Names = []
+    callback_Display_Names = []
+    callback_Definition = []
+
+    for callback in list_Callbacks:
+        callback_Names.append(callback.getName())
+        callback_Display_Names.append(callback.getDisplayName())
+        callback_Definition.append(callback.getDefinition())
+    g.callback_Names = callback_Names
+    g.callback_Display_Names = callback_Display_Names
+    g.callback_Info =  callback_Definition
+    g.callbacks = zip(g.callback_Names, g.callback_Display_Names, g.callback_Info)
 
 @bp.route("/new_experiment")
 def new_experiment():
@@ -337,7 +352,8 @@ def calculate():
 
 
 @bp.route("/run_experiment", methods=["POST"])
-def run_experiment():
+def run_experiment():    
+
     processes_input = request.form.get("processes")
     data = json.loads(processes_input)
 
@@ -347,41 +363,70 @@ def run_experiment():
     data["csvFileName"] = filename
     data["csvFile"] = csv_file
 
-    # Split
-    if data['validation'] == "Split":
-        status, msg, Metrics = Validation.Split(data)
+    # Perform MLA if chosen
+    if data['methodology'] == "MLA":
+        # Split
+        if data['validation'] == "Split":
+            status, msg, Metrics = MLA_Validation.Split(data)
 
-    # K-Fold
-    elif data['validation'] == "K-Fold":
-        status, msg, Metrics = Validation.K_Fold(data)
+        # K-Fold
+        elif data['validation'] == "K-Fold":
+            status, msg, Metrics = MLA_Validation.K_Fold(data)
 
-    if status == "worked":
-        # Open json file for the experiment.
-        baseFolder = os.getcwd()
-        locationSavedResults = Path(baseFolder) / "SavedResults"
-        filename = secure_filename(data["projectName"] + ".json")
-        filepath = locationSavedResults / filename.lower()
-        if os.path.exists(filepath):
-            os.remove(filepath)
-            fp = open(filepath, "a")
-        else:
-            fp = open(filepath, "a")
+        if status == "worked":
+            # Open json file for the experiment.
+            baseFolder = os.getcwd()
+            locationSavedResults = Path(baseFolder) / "SavedResults"
+            filename = secure_filename(data["projectName"] + ".json")
+            filepath = locationSavedResults / filename.lower()
+            if os.path.exists(filepath):
+                os.remove(filepath)
+                fp = open(filepath, "a")
+            else:
+                fp = open(filepath, "a")
+    
+            # write to json file
+            metrics_Dump = json.dumps(Metrics)
 
-        # write to json file
-        metrics_Dump = json.dumps(Metrics)
+            fp.write(metrics_Dump)
 
-        fp.write(metrics_Dump)
+            # close the connection
+            fp.close()
 
-        # close the connection
-        fp.close()
+        Results = [status, msg, Metrics]
 
-    Results = [status, msg, Metrics]
+    # Perform Deep Learning Neural Network.
+    if data['methodology'] == "DLANN":
+        # Split
+        status, msg, Metrics = DLANN_Validation.Split(data)
+
+        if status == "worked":
+            # Open json file for the experiment.
+            baseFolder = os.getcwd()
+            locationSavedResults = Path(baseFolder) / "SavedResults"
+            filename = secure_filename(data["projectName"] + ".json")
+            filepath = locationSavedResults / filename.lower()
+            if os.path.exists(filepath):
+                os.remove(filepath)
+                fp = open(filepath, "a")
+            else:
+                fp = open(filepath, "a")
+    
+            # write to json file
+            metrics_Dump = json.dumps(Metrics)
+
+            fp.write(metrics_Dump)
+
+            # close the connection
+            fp.close()
+
+        Results = [status, msg, Metrics]
 
     return json.dumps(Results)
 
 
 @bp.route("/getResults", methods=["POST"])
-def getResults():
+def getResults(): 
     output = request.get_json()
     formData = json.loads(output)
 
@@ -397,7 +442,7 @@ def getResults():
 
     # close the connection
     fp.close()
-
+    
     return json.dumps(Metrics)
 
 
@@ -410,7 +455,6 @@ def getAlgorithmParameters():
 
     return json.dumps(Parameters)
 
-
 @bp.route("/getCategoryPreopts", methods=["POST"])
 def getCategoryPreopts():
     output = request.get_json()
@@ -419,6 +463,15 @@ def getCategoryPreopts():
     preopts = Preoptimization.getCategoryPreopts(data["Category"])
 
     return json.dumps(preopts)
+
+@bp.route("/getCategoryLayers", methods=["POST"])
+def getCategoryLayers():
+    output = request.get_json()
+    data = json.loads(output)
+
+    layers = NeuralNetwork.getCategoryLayers(data["Category"])
+
+    return json.dumps(layers)
 
 
 @bp.route("/results")
@@ -443,6 +496,16 @@ def getCSVResults():
 
     return json.dumps(Results)
 
+@bp.route("/getModelSummary", methods=["POST"])
+def getModelSummary():
+    processes_input = request.form.get("processes")
+    data = json.loads(processes_input)
+
+    status, msg, info = NeuralNetwork.getModelSummary(data)
+
+    Results = [status, msg, info]
+
+    return json.dumps(Results)
 
 @bp.route("/downloadCSV", methods=["POST"])
 def downloadCSV():
@@ -460,7 +523,6 @@ def downloadCSV():
     Results = [status, msg, info]
 
     return json.dumps(Results)
-
 
 @bp.route("/getPreoptParameters", methods=["POST"])
 def getPreoptParameters():
@@ -481,9 +543,16 @@ def getLayerParameters():
 
     return json.dumps(Parameters)
 
+@bp.route("/getCallbackParameters", methods=["Post"])
+def getCallbackParameters():
+    output = request.get_json()
+    data = json.loads(output)
+
+    Parameters = Callbacks.getParameters(data["Callback"])
+
+    return json.dumps(Parameters)
+
 # Gets the colummn names inside the csv.
-
-
 @bp.route("/getCSVColumnTitles", methods=["POST"])
 def getCSVColumnTitles():
     filename = request.form.get("csvFileName")
@@ -494,3 +563,17 @@ def getCSVColumnTitles():
     columnTitles = Preoptimization.getCSVColumnTitles(data)
 
     return json.dumps(columnTitles)
+
+# Gets all the options that is available via the keras model.compile 
+@bp.route("/getCompilerOptions", methods=["POST"])
+def getCompilerOptions():
+    compiler_Options = Compiler.getCompilerOptions()
+
+    return json.dumps(compiler_Options)
+
+# Gets all the options that is available for the model to fit,train, and validate.
+@bp.route("/getValidationOptions", methods=["POST"])
+def getValidationOptions():
+    validation_Options = NN_Validation.getValidationOptions()
+
+    return json.dumps(validation_Options)
